More work on processors and now testing
